+++ djangoapp/financeiro/urls.py
@@ -12,6 +12,6 @@ urlpatterns = [
 
     path('concilia-contas/', views.concilia_contas_view, name='concilia_contas'),
     # path('concilia-contas/validar/', views.validar_ofx, name='validar_ofx'),
-    # path('concilia-contas/incluir/', views.incluir_conta_ofx, name='incluir_conta_ofx'),
+    path('concilia/incluir/', views.incluir_conta_conciliacao, name='incluir_conta_conciliacao'),
 
 ]